Storage for plugin data: hermione now has a leveldb brain.
If you provide a brain.dbpath option string, a leveldb instance
is created. Messages are decorated with a `brain` object that is
a sublevel-namespaced db.  Plugins thus get a sublevel-namespaced
db instance to play with as they choose. The value encoding is
`json` so plugins should pass js objects as values to put().