Major refactoring of how the server is built & kicked off.
You can now run the bot using the script `bin/opsbot.js`. This takes an
optional command-line argument to locate the config file. It loads the
config & passes it on to the restify server, which is now wrapped in
the Opbot object.

Wrote a few tests for the server itself. Code coverage is now over 80%.